# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/users": {
      "post": {
        "operationId": "UserController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuario creado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Datos de entrada inválidos"
          },
          "409": {
            "description": "El email ya está registrado"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Crear un nuevo usuario",
        "tags": [
          "users"
        ]
      },
      "get": {
        "operationId": "UserController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Lista de usuarios obtenida exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Obtener todos los usuarios",
        "tags": [
          "users"
        ]
      }
    },
    "/users/{id}": {
      "get": {
        "operationId": "UserController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID del usuario",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Usuario encontrado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Usuario no encontrado"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Obtener un usuario por ID",
        "tags": [
          "users"
        ]
      },
      "patch": {
        "operationId": "UserController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID del usuario",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Usuario actualizado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Usuario no encontrado"
          },
          "409": {
            "description": "El email ya está registrado"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Actualizar un usuario",
        "tags": [
          "users"
        ]
      },
      "delete": {
        "operationId": "UserController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID del usuario",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Usuario eliminado exitosamente"
          },
          "404": {
            "description": "Usuario no encontrado"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Eliminar un usuario",
        "tags": [
          "users"
        ]
      }
    },
    "/users/profile/me": {
      "get": {
        "operationId": "UserController_getProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Perfil del usuario obtenido exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "No autorizado"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Obtener perfil del usuario autenticado",
        "tags": [
          "users"
        ]
      },
      "patch": {
        "operationId": "UserController_updateProfile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Perfil actualizado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "No autorizado"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Actualizar perfil del usuario autenticado",
        "tags": [
          "users"
        ]
      }
    },
...
  },
  "info": {
    "title": "Salon Backend API",
    "description": "API completa para el sistema de salon con autenticacion JWT",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
...
      "UserResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID único del usuario",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "email": {
            "type": "string",
            "description": "Email del usuario",
            "example": "usuario@ejemplo.com"
          },
          "firstName": {
            "type": "string",
            "description": "Nombre del usuario",
            "example": "Juan"
          },
          "lastName": {
            "type": "string",
            "description": "Apellido del usuario",
            "example": "Pérez"
          },
          "fullName": {
            "type": "string",
            "description": "Nombre completo del usuario",
            "example": "Juan Pérez"
          },
          "phone": {
            "type": "string",
            "description": "Número de teléfono del usuario",
            "example": "+1234567890",
            "nullable": true
          },
          "isActive": {
            "type": "boolean",
            "description": "Estado activo del usuario",
            "example": true
          },
          "emailVerified": {
            "type": "boolean",
            "description": "Estado de verificación del email",
            "example": false
          },
          "role": {
            "type": "string",
            "description": "Rol del usuario en el sistema",
            "enum": [
              "admin",
              "cliente",
              "prestador_servicio"
            ],
            "example": "cliente"
          },
          "providerType": {
            "type": "string",
            "description": "Tipo de prestador",
            "enum": [
              "barbero",
              "estilista",
              "manicurista",
              "maquilladora"
            ],
            "example": "barbero",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Fecha de creación del usuario",
            "example": "2024-01-01T00:00:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Fecha de última actualización del usuario",
            "example": "2024-01-01T00:00:00.000Z"
          }
        },
        "required": [
          "id",
          "email",
          "firstName",
          "lastName",
          "fullName",
          "phone",
          "isActive",
          "emailVerified",
          "role",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Email del usuario",
            "example": "usuario@ejemplo.com"
          },
          "password": {
            "type": "string",
            "description": "Contraseña del usuario",
            "example": "MiPassword123!",
            "minLength": 8
          },
          "firstName": {
            "type": "string",
            "description": "Nombre del usuario",
            "example": "Juan"
          },
          "lastName": {
            "type": "string",
            "description": "Apellido del usuario",
            "example": "Pérez"
          },
          "phone": {
            "type": "string",
            "description": "Número de teléfono del usuario",
            "example": "+1234567890"
          },
          "role": {
            "type": "string",
            "description": "Rol del usuario en el sistema",
            "enum": [
              "admin",
              "cliente",
              "prestador_servicio"
            ],
            "example": "cliente",
            "default": "cliente"
          },
          "providerType": {
            "type": "string",
            "description": "Tipo de prestador (obligatorio para rol prestador_servicio)",
            "enum": [
              "barbero",
              "estilista",
              "manicurista",
              "maquilladora"
            ],
            "example": "barbero"
          }
        },
        "required": [
          "email",
          "password",
          "firstName",
          "lastName"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Email del usuario",
            "example": "usuario@ejemplo.com"
          },
          "firstName": {
            "type": "string",
            "description": "Nombre del usuario",
            "example": "Juan"
          },
          "lastName": {
            "type": "string",
            "description": "Apellido del usuario",
            "example": "Pérez"
          },
          "phone": {
            "type": "string",
            "description": "Número de teléfono del usuario",
            "example": "+1234567890"
          },
          "role": {
            "type": "string",
            "description": "Rol del usuario en el sistema",
            "enum": [
              "admin",
              "cliente",
              "prestador_servicio"
            ],
            "example": "cliente",
            "default": "cliente"
          },
          "providerType": {
            "type": "string",
            "description": "Tipo de prestador (obligatorio para rol prestador_servicio)",
            "enum": [
              "barbero",
              "estilista",
              "manicurista",
              "maquilladora"
            ],
            "example": "barbero"
          },
          "isActive": {
            "type": "boolean",
            "description": "Estado activo del usuario",
            "example": true
          },
          "emailVerificationToken": {
            "type": "string",
            "description": "Token de verificación de email",
            "example": "abc123def456"
          }
        }
      },
...
    }
  }
}